+++ src/shared/hooks/challenge/useApprovalChallenge.ts
@@ -4,12 +4,12 @@ import { toast } from 'sonner';
 import { approvalStudentChallengeResult } from '@/shared/apis/challenge';
 import { queryKeys } from '@/shared/constants/query-keys';
 
-function useApprovalChallenge() {
+function useApprovalChallenge({ academyId, releasedChallengeId }: { academyId: number; releasedChallengeId: number }) {
   const queryClient = useQueryClient();
   return useMutation({
     mutationFn: approvalStudentChallengeResult,
     onSuccess: async () => {
-      await queryClient.invalidateQueries({ queryKey: [queryKeys.CHALLENGE_STUDENT_PROGRESS_LIST] });
+      await queryClient.invalidateQueries({ queryKey: [queryKeys.CHALLENGE_STUDENT_PROGRESS_LIST, { academyId }, { releasedChallengeId }] });
       toast.success('승인 변경 완료');
     },
     onError: (error) => {
